backend/requirements.txt
jinja2
Flask-PyMongo
watchdog
pyqtwebengine<5.13
pyqt5==5.12.3
pymongo